New pytest cicd

+++ test/test_get_new_data_from_database.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/env bash
-
+ 
 cat > test_database.sql <<EOM
 \c test_database
 
@@ -26,11 +26,11 @@ EOM
 
 cat > test/test_database.ini << EOM
 [postgresql_test_database]
-host=localhost
+host=$PGHOST
 port=5432
 database=test_database
 user=$USER
-password=password
+password=$PGPASSWORD
 EOM
 
 psql -c 'DROP DATABASE IF EXISTS test_database;'
@@ -38,4 +38,4 @@ psql -c 'CREATE DATABASE test_database;'
 psql -f test_database.sql
 rm test_database.sql
 pytest test/test_get_new_data_from_database.py -vvvrP
-psql -c "DROP DATABASE IF EXISTS test_database"
+if  [[ $PGUSER!='postgres_user' ]]; then psql -c "DROP DATABASE IF EXISTS test_database"; fi
